requirements.txt
absl-py==2.1.0
accelerate==0.23.0
action-msgs==2.0.1
action-tutorials-interfaces==0.33.1
action-tutorials-py==0.33.1
actionlib-msgs==5.3.0
addict==2.4.0
aiofiles==23.2.1
altair==5.5.0
ament-cmake-test==2.3.2
ament-copyright==0.16.3
ament-cppcheck==0.16.3
ament-cpplint==0.16.3
ament-flake8==0.16.3
ament-index-python==1.7.0
ament-lint==0.16.3
ament-lint-cmake==0.16.3
ament-package==0.16.3
ament-pep257==0.16.3
ament-uncrustify==0.16.3
ament-xmllint==0.16.3
angles==1.16.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
appdirs==1.4.4
apriltag==0.0.16
asttokens==2.4.1
attrs==25.1.0
black==24.10.0
blinker==1.9.0
bosdyn-api==4.1.1
bosdyn-choreography-client==4.1.1
bosdyn-client==4.1.1
bosdyn-core==4.1.1
bosdyn-mission==4.1.1
bosdyn-orbit==4.1.1
builtin-interfaces==2.0.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cityscapesScripts==2.2.2
click==8.1.8
cloudpickle==3.1.1
coloredlogs==15.0.1
comm==0.2.2
composition-interfaces==2.0.1
ConfigArgParse==1.7
contourpy==1.3.1
cv-bridge==3.4.0
cycler==0.12.1
Cython==3.0.2
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
deepspeed==0.10.3
demo-nodes-py==0.33.1
Deprecated==1.2.18
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@9604f5995cc628619f0e4fd913453b4d7d61db3f
diagnostic-msgs==5.3.0
diagnostic-updater==3.1.2
diffdist==0.1
docker-pycreds==0.4.0
domain-coordinator==0.12.0
einops==0.7.0
example-interfaces==0.12.0
examples-rclpy-executors==0.19.1
examples-rclpy-minimal-action-client==0.19.1
examples-rclpy-minimal-action-server==0.19.1
examples-rclpy-minimal-client==0.19.1
examples-rclpy-minimal-publisher==0.19.1
examples-rclpy-minimal-service==0.19.1
examples-rclpy-minimal-subscriber==0.19.1
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.7
fastjsonschema==2.20.0
ffmpeg==1.4
ffmpy==0.5.0
filelock==3.17.0
Flask==3.0.3
flatbuffers==25.1.24
fonttools==4.55.6
fsspec==2024.12.0
ftfy==6.1.1
fvcore==0.1.5.post20221221
geometry-msgs==5.3.0
gitdb==4.0.12
GitPython==3.1.44
gradio==3.42.0
gradio_client==0.5.0
grpcio==1.70.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.17.3
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
image-geometry==3.4.0
imageio==2.36.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
infinibatch==0.1.1
interactive-markers==2.5.3
iopath==0.1.9
ipython==8.29.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
json-tricks==3.17.3
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
kornia==0.6.8
laser-geometry==2.7.0
launch==3.4.0
launch-pytest==3.4.0
launch-ros==0.26.4
launch-testing==3.4.0
launch-testing-ros==0.26.4
launch-xml==3.4.0
launch-yaml==3.4.0
lazy_loader==0.4
lifecycle-msgs==2.0.1
llvmlite==0.43.0
logging-demo==0.33.1
map-msgs==2.4.0
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mediapipe==0.10.20
message-filters==4.11.0
ml_dtypes==0.5.1
more-itertools==10.6.0
mpi4py==4.0.1
mpmath==1.3.0
mup==1.0.0
mypy-extensions==1.0.0
narwhals==1.24.0
nav-msgs==5.3.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
nltk==3.8.1
numba==0.60.0
numpy==1.23.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
open3d==0.18.0
openai-whisper==20240930
opencv-contrib-python==4.11.0.86
opencv-python==4.8.1.78
opt_einsum==3.4.0
orjson==3.10.15
osrf-pycommon==2.1.4
packaging==24.2
pandas==2.0.3
parso==0.8.4
pathspec==0.12.1
pathtools==0.1.2
pcl-msgs==1.0.0
pendulum-msgs==0.33.1
pexpect==4.9.0
Pillow==9.4.0
platformdirs==4.3.6
plotly==5.24.1
portalocker==3.1.1
prompt_toolkit==3.0.48
protobuf==4.25.6
psutil==6.1.1
ptyprocess==0.7.0
pupil-apriltags==1.0.4.post10
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==13.0.0
pycocotools==2.0.7
pycparser==2.22
pydantic==1.10.21
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pynmea2==1.19.0
pyparsing==3.2.1
pyquaternion==0.9.9
pyrealsense2==2.54.2.5684
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-qt-binding==2.1.1
pytz==2024.2
PyWavelets==1.8.0
PyYAML==6.0.1
qt-dotgraph==2.7.2
qt-gui==2.7.2
qt-gui-cpp==2.7.2
qt-gui-py-common==2.7.2
quality-of-service-demo-py==0.33.1
rcl-interfaces==2.0.1
rclpy==7.0.1
rcutils==6.5.2
referencing==0.35.1
regex==2023.10.3
requests==2.32.3
resource-retriever==3.4.0
retrying==1.3.4
rmw-dds-common==3.0.0
ros2action==0.31.1
ros2bag==0.24.0
ros2bag-mcap-cli==0.24.0
ros2bag-sqlite3-cli==0.24.0
ros2cli==0.31.1
ros2component==0.31.1
ros2doctor==0.31.1
ros2interface==0.31.1
ros2launch==0.26.4
ros2lifecycle==0.31.1
ros2multicast==0.31.1
ros2node==0.31.1
ros2param==0.31.1
ros2pkg==0.31.1
ros2run==0.31.1
ros2service==0.31.1
ros2topic==0.31.1
rosbag2-interfaces==0.24.0
rosbag2-py==0.24.0
rosgraph-msgs==2.0.1
rosidl-adapter==4.5.1
rosidl-cli==4.5.1
rosidl-cmake==4.5.1
rosidl-generator-c==4.5.1
rosidl-generator-cpp==4.5.1
rosidl-generator-py==0.21.1
rosidl-generator-type-description==4.5.1
rosidl-parser==4.5.1
rosidl-pycommon==4.5.1
rosidl-runtime-py==0.13.0
rosidl-typesupport-c==3.2.0
rosidl-typesupport-cpp==3.2.0
rosidl-typesupport-fastrtps-c==3.4.0
rosidl-typesupport-fastrtps-cpp==3.4.0
rosidl-typesupport-introspection-c==4.5.1
rosidl-typesupport-introspection-cpp==4.5.1
rpds-py==0.21.0
rpyutils==0.4.1
rqt-action==2.2.0
rqt-bag==1.5.1
rqt-bag-plugins==1.5.1
rqt-console==2.2.1
rqt-graph==1.5.2
rqt-gui==1.5.0
rqt-gui-py==1.5.0
rqt-msg==1.5.1
rqt-plot==1.3.2
rqt-publisher==1.7.1
rqt-py-common==1.5.0
rqt-py-console==1.2.2
rqt-reconfigure==1.6.2
rqt-service-caller==1.2.1
rqt-shell==1.2.1
rqt-srv==1.2.2
rqt-topic==1.7.1
safetensors==0.5.2
scikit-image==0.21.0
scikit-learn==1.3.1
scikit-spatial==8.1.0
scipy==1.13.1
seaborn==0.13.2
SEEM @ git+https://github.com/UX-Decoder/Segment-Everything-Everywhere-All-At-Once.git@5feca7f4b600e5b79ddfd38c25592181f413b817
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
Semantic-SAM @ git+https://github.com/UX-Decoder/Semantic-SAM.git@b6a154c44a4e64ed7a70cd7fefcb131acb6b4aad
semantic-version==2.10.0
sensor-msgs==5.3.0
sensor-msgs-py==5.3.0
sentencepiece==0.1.99
sentry-sdk==2.20.0
service-msgs==2.0.1
setproctitle==1.3.4
shape-msgs==5.3.0
Shapely==1.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
sros2==0.13.0
stack-data==0.6.3
starlette==0.45.3
statistics-msgs==2.0.1
std-msgs==5.3.0
std-srvs==5.3.0
stereo-msgs==5.3.0
sympy==1.13.1
tabulate==0.9.0
teleop-twist-keyboard==2.3.2
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
tf2-geometry-msgs==0.36.0
tf2-kdl==0.36.0
tf2-msgs==0.36.0
tf2-py==0.36.0
tf2-ros-py==0.36.0
tf2-sensor-msgs==0.36.0
tf2-tools==0.36.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tiktoken==0.8.0
timm==0.4.12
tokenizers==0.14.1
toml==0.10.2
tomli==2.0.1
topic-monitor==0.33.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
traitlets==5.14.3
trajectory-msgs==5.3.0
transformers==4.34.0
triton==3.1.0
turtlesim==1.8.1
type-description-interfaces==2.0.1
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.2
unique-identifier-msgs==2.5.0
urllib3==2.3.0
uvicorn==0.34.0
vision-datasets==0.2.2
visualization-msgs==5.3.0
wandb==0.15.12
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wrapt==1.17.2
xacro==2.0.10
yacs==0.1.8
zipp==3.21.0
